+++ src/os2datascanner/projects/admin/adminapp/models/__init__.py
@@ -3,3 +3,4 @@ from . import apikey  # noqa
 from . import rules  # noqa
 from . import scannerjobs  # noqa
 from . import sensitivity_level  # noqa
+from . import MSGraphSharePointSite  # noqa